+++ src/views/Dashboard/DashboardV2.js
@@ -55,7 +55,7 @@ import avalanche24Icon from "../../img/ic_avalanche_24.svg";
 
 import AssetDropdown from "./AssetDropdown";
 import SEO from "../../components/Common/SEO";
-import TooltipCard, { TooltipCardRow } from "./TooltipCard";
+import TooltipCard, { TooltipCardRow } from "../../components/Tooltip/TooltipCard";
 import useTotalVolume from "../../hooks/useTotalVolume";
 const ACTIVE_CHAIN_IDS = [ARBITRUM, AVALANCHE];
